add some

+++ dijkstra.h
@@ -9,10 +9,10 @@
 template <typename T>
 struct Result {
   std::vector<std::optional<T>> dist;
-  std::vector<std::optional<int32_t>> prev;
-  std::vector<int32_t> Path(int32_t dest) const {
+  std::vector<std::optional<int64_t>> prev;
+  std::vector<int64_t> Path(int64_t dest) const {
     dassert(prev[dest]);
-    std::vector<int32_t> v;
+    std::vector<int64_t> v;
     while (dest >= 0) {
       v.push_back(dest);
       dest = *prev[dest];
@@ -25,17 +25,17 @@ struct Result {
 template <typename T>
 struct DijkstraState {
   T dist;
-  int32_t node;
-  int32_t prev;
+  int64_t node;
+  int64_t prev;
   bool operator<(const DijkstraState& s) const { return dist > s.dist; }
 };
 
 template <typename T>
-Result<T> Dijkstra(const WeightedGraph<T>& graph, int32_t source) {
+Result<T> Dijkstra(const WeightedGraph<T>& graph, int64_t source) {
   const int32_t n = graph.size();
 
   std::vector<std::optional<T>> dist(n);
-  std::vector<std::optional<int32_t>> prev(n);
+  std::vector<std::optional<int64_t>> prev(n);
 
   std::priority_queue<DijkstraState<T>> que;
   que.push({0, source, -1});
